Did more work on ScSplash class (found some design issues)
ScSplash class mostly done but issue of getting engine level info into scenes has shown up.
For example, I need to see the delta time but can't figure out a way to pass it to the scene without literally passing it each frame.
Ideally, I should be able to say something like "Engine.delta" or "Time.delta"  in the derived class and it should just give me the delta time.
Adding a reference to the game class causes cyclic dependencies but otherwise I will have to have the engine send over all the possible data I would need every frame.
Even if it's not that expensive (only dozens of variables max), it is really annoying. Need to figure this out as soon as possible.